requirements.txt
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
scipy==1.11.4
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.5
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dateutil==2.8.2
tqdm==4.66.1
Pillow==10.1.0
plotly==5.18.0
